Added Tmp adn $bitforbit slit method

+++ powershell/zip_maker.ps1
@@ -9,6 +9,9 @@
     [Parameter(Mandatory=$true)]
     [int32]$maxsize,
 
+    [Parameter(Mandatory=$false)]
+    [int32]$bitforbitsplit,
+
     [switch]$debugger
 )
 
@@ -26,25 +29,28 @@ Remove-Item $PSScriptRoot\tmp\maxlen.ini -Force -ErrorAction SilentlyContinue
 
 $array = @()
 
+if($bitforbitsplit -eq 1){
+    Compress-Archive -DestinationPath $outpath -Path $path -CompressionLevel Optimal -Confirm
 
-#debug
-#$path = D:\School\English
+}else{
 
-$FileList = Get-ChildItem -LiteralPath $path -File
 
-$Results = foreach ($FL_Item in $FileList)
-    {
-    [PSCustomObject]@{
-        Name = $FL_Item.Name
-        Size_KB = '{0,7:N2}' -f ($FL_Item.Length / 1KB)
+
+    $FileList = Get-ChildItem -LiteralPath $path -File
+
+    $Results = foreach ($FL_Item in $FileList)
+        {
+        [PSCustomObject]@{
+            Name = $FL_Item.Name
+            Size_KB = '{0,7:N2}' -f ($FL_Item.Length / 1KB)
+            }
         }
+
+    Remove-Item -Path \CSV\sizes.csv -Force -ErrorAction SilentlyContinue
+    echo $Results
+
+    $Results | Export-Csv -Confirm -Force -Delimiter % -Path $PSScriptRoot\CSV\results.csv
+    $maxsize | Out-File  -Path $PSScriptRoot\tmp\maxlen.ini -Force
     }
-
-Remove-Item -Path \CSV\sizes.csv -Force -ErrorAction SilentlyContinue
-echo $Results
-
-$Results | Export-Csv -Confirm -Force -Delimiter % -Path $PSScriptRoot\CSV\results.csv
-$maxsize | Out-File  -Path $PSScriptRoot\tmp\maxlen.ini -Force
-
